+++ CHANGELOG.md
@@ -18,6 +18,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v2.24.1 - 2021-03-07
+
+### Changed
+
+- 予約検索の管理ユーザー条件をreservation.brokerに割り当てるように調整
+
 ## v2.24.0 - 2021-03-03
 
 ### Added